performs all matrix operations

+++ matrices/kernels.cuh
@@ -14,11 +14,11 @@
 
 using namespace std;
 
-void matrixMultiplication(float *A, float *B, float *C, int N);
+void matrixMultiplication(double *A, double *B, double *C, int N);
 
-void matrixAddition(float *A, float *B, float *C, int N);
+void matrixAddition(double *A, double *B, double *C, int N);
 
-void matrixSubtraction(float *A, float *B, float *C, int N);
+void matrixSubtraction(double *A, double *B, double *C, int N);
 
 void matrixOperations(GtkWidget *widget, gpointer data);
